third_party/WebKit/LayoutTests/typedcssom/cssSkew.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>
var EPSILON = 1e-6; // float epsilon

function angle(deg) { return new CSSUnitValue(deg, 'deg'); }

function tanDegrees(degrees) {
  var radians = degrees * Math.PI / 180;
  return Math.tan(radians);
}

function assert_array_approx_equals(actual, expected) {
  for (var i = 0; i < actual.length; i++) {
    assert_approx_equals(actual[i], expected[i], EPSILON);
  }
}

function assert_matrix_approx_equals(actual, expected) {
  assert_array_approx_equals(actual.toFloat64Array(), expected.toFloat64Array());
}

var values = [
  {input: new CSSSkew(angle(0), angle(0)), ax: 0, ay: 0, cssText: "skew(0deg, 0deg)"},
  {input: new CSSSkew(angle(1), angle(2)), ax: 1, ay: 2, cssText: "skew(1deg, 2deg)"},
  {input: new CSSSkew(angle(-2), angle(-4)), ax: -2, ay: -4, cssText: "skew(-2deg, -4deg)"},
  {input: new CSSSkew(angle(3.4), angle(2.7)), ax: 3.4, ay: 2.7, cssText: "skew(3.4deg, 2.7deg)"},
  {input: new CSSSkew(new CSSUnitValue(1, 'rad'), angle(0)), ax: 57.2957795, ay: 0, cssText: "skew(1rad, 0deg)"},
  {input: new CSSSkew(angle(0), new CSSUnitValue(1, 'rad')), ax: 0, ay: 57.2957795, cssText: "skew(0deg, 1rad)"}
];

test(function() {
  for (var i = 0; i < values.length; ++i) {
    assert_approx_equals(values[i].input.ax.degrees, values[i].ax, EPSILON);
    assert_approx_equals(values[i].input.ay.degrees, values[i].ay, EPSILON);
  }
}, "(ax, ay) values for CSSSkew are correct.");

test(function() {
  for (var i = 0; i < values.length; ++i) {
    assert_true(values[i].input.is2D());
  }
}, "is2D values for CSSSkew are correct.");

test(function() {
  for (var i = 0; i < values.length; ++i) {
    assert_equals(values[i].input.toString(), values[i].cssText);
  }
}, "toString for CSSSkew is correct.");

test(function() {
  assert_throws(new TypeError(), function() { new CSSSkew(); });
  assert_throws(new TypeError(), function() { new CSSSkew(null); });
  assert_throws(new TypeError(), function() { new CSSSkew(1); });
  assert_throws(new TypeError(), function() { new CSSSkew('1'); });
  assert_throws(new TypeError(), function() { new CSSSkew(angle(1)); });
}, "Invalid arguments for CSSSkew throws an exception.");

test(function() {
  for (var i = 0; i < values.length; ++i) {
    var input = values[i].input;
    var inputAsMatrix = input.asMatrix();
    assert_true(inputAsMatrix.is2D());
    var tanAx = tanDegrees(input.ax.degrees);
    var tanAy = tanDegrees(input.ay.degrees);
    var expectedMatrix = new CSSMatrixComponent(new DOMMatrixReadOnly([1, tanAy, tanAx, 1, 0, 0]));
    for (var attribute in expectedMatrix) {
      if (attribute == "matrix") {
        assert_matrix_approx_equals(inputAsMatrix[attribute], expectedMatrix[attribute]);
      } else {
        assert_equals(inputAsMatrix[attribute], expectedMatrix[attribute]);
      }
    }
  }
}, "asMatrix is constructed correctly for CSSSkew.");

test(function() {
  var skew = new CSSSkew(new CSSUnitValue(1, 'deg'), new CSSUnitValue(2, 'deg'));
  skew.ax = new CSSUnitValue(3, 'deg');
  skew.ay = new CSSUnitValue(3, 'rad');

  assert_equals(skew.ax.value, 3);
  assert_equals(skew.ay.value, 3);
  assert_equals(skew.ax.unit, 'deg');
  assert_equals(skew.ay.unit, 'rad');
}, "Setting ax and ay for CSSSkew with valid CSSUnitValues");

for (let a of ['ax', 'ay']) {
  test(() => {
    var skew = new CSSSkew(
        new CSSUnitValue(1, 'deg'), new CSSUnitValue(2, 'deg'));
    assert_throws(new TypeError(), () => {
      skew[a] = new CSSUnitValue(1, 'px');
    });
    assert_throws(new TypeError(), () => {
      skew[a] = 'bananas';
    });
    assert_throws(new TypeError(), () => {
      skew[a] = null;
    });
  }, "Setting " + a + " with invalid values");
}

</script>
